Move everest_config_with_defaults to utils
