+++ tests/everest/test_config_file_loader.py
@@ -7,7 +7,7 @@ import pytest
 from everest import config_file_loader as loader
 from everest.config import EverestConfig
 from everest.config.everest_config import EverestValidationError
-from tests.everest.conftest import everest_config_with_defaults
+from tests.everest.utils import everest_config_with_defaults
 
 
 def test_load_yaml(tmp_path):
