+++ tests/everest/test_controls.py
@@ -8,7 +8,7 @@ from everest.config.control_variable_config import (
     ControlVariableConfig,
 )
 from everest.optimizer.everest2ropt import everest2ropt
-from tests.everest.conftest import everest_config_with_defaults
+from tests.everest.utils import everest_config_with_defaults
 
 
 def test_that_duplicate_control_group_name_is_invalid(min_config):
