+++ tests/everest/test_domain_transforms.py
@@ -4,7 +4,7 @@ import pytest
 from everest.config import EverestConfig
 from everest.optimizer.everest2ropt import everest2ropt
 from everest.optimizer.opt_model_transforms import get_optimization_domain_transforms
-from tests.everest.conftest import everest_config_with_defaults
+from tests.everest.utils import everest_config_with_defaults
 
 
 @pytest.fixture
