+++ tests/everest/test_environment.py
@@ -8,7 +8,7 @@ from ert.base_model_context import use_runtime_plugins
 from ert.plugins import get_site_plugins
 from ert.run_models.everest_run_model import EverestRunModel
 from everest.config import EverestConfig
-from tests.everest.conftest import everest_config_with_defaults
+from tests.everest.utils import everest_config_with_defaults
 
 
 def test_that_default_everestconfig_lints():
