+++ tests/everest/test_output_constraints.py
@@ -6,7 +6,7 @@ from ert.plugins import get_site_plugins
 from ert.run_models.everest_run_model import EverestRunModel
 from everest.config import EverestConfig, OptimizationConfig, OutputConstraintConfig
 from everest.optimizer.everest2ropt import everest2ropt
-from tests.everest.conftest import everest_config_with_defaults
+from tests.everest.utils import everest_config_with_defaults
 
 
 def test_constraints_init(tmp_path):
