+++ tests/everest/utils/__init__.py
@@ -1,7 +1,9 @@
 from .optimal_result import get_optimal_result
 from .utils import (
+    MIN_CONFIG,
     MockParser,
     capture_streams,
+    everest_config_with_defaults,
     relpath,
     satisfy,
     satisfy_callable,
@@ -10,8 +12,10 @@ from .utils import (
 )
 
 __all__ = [
+    "MIN_CONFIG",
     "MockParser",
     "capture_streams",
+    "everest_config_with_defaults",
     "get_optimal_result",
     "relpath",
     "satisfy",
